Fixed some possible bugs for CSFileText. Improved file I/O error reporting. Moved CSFileText class from CSFile.h/cpp to CSFileText.h/cpp. Now CFile, CSFile, CSFileText and child classes support int file offsets. Before, the data type was size_t, but the functions used internally to seek/read/etc used int/long/dword/off_t types, so the behavior was inconsistent between operating systems, architecture and even between CFile/CSFile and CSFileText and child classes (like CacheableScriptFile). If we'll every need support for files bigger than 2 GB, we should review the whole implementation to use other internal functions.

+++ src/common/CResourceBase.h
@@ -534,7 +534,7 @@ public:
 
 	bool IsLinked() const;	// been loaded from the scripts ?
 	CResourceScript * GetLinkFile() const;
-	size_t GetLinkOffset() const;
+	int GetLinkOffset() const;
 	void SetLink( CResourceScript * pScript );
     void CopyTransfer( CResourceLink * pLink );
 	void ScanSection( RES_TYPE restype );
